Add support for HSPI peripheral on ESP32 with WS2801
Bitbanging the WS2801 protocol is very flickery. Cause ESP32 has a
second SPI peripheral, we can just use that for the second LED strip,
avoiding the need for bitbanging if you have two strips or less.

Only adding it for WS2801 because it is the only LED strip I have and
can test with.